fix: remove the Investments section from the Net Worth page
- Remove `NetWorth.investments` and stop computing holdings in `ledgerNetWorth()`
- Move the holdings grid's column sizes and width into `investments-columns.ts`
- Delete the `InvestmentsSection` wrapper, now unused by the Net Worth page
- Update the Net Worth and Investments tests to match
