+++ packages/desktop/src/renderer/components/accountant24/__tests__/investments-columns.test.ts
@@ -3,13 +3,21 @@
 // Spec for the Investments view's table config: its storage key and
 // defaults over the shared core — only Cost, P&L, and Allocation toggle,
 // hidden by default; Commodity, Quantity, Price, and Value are the page's
-// spine. Widths reuse the Net Worth page's holdings model (already
-// specified in net-worth-columns.test.ts).
+// spine. Also the static width model: fixed defaults whose default-visible
+// set spans the same width as the Net Worth account tables, optional
+// columns growing the table past the floor, and minimums that keep every
+// header pill readable.
 
 import { beforeAll, beforeEach, describe, expect, it } from "vitest";
 import { installJsdomPolyfills } from "@/test/jsdomPolyfills";
-import { INVESTMENTS_TABLE_KEY, loadInvestmentsTableConfig, saveInvestmentsTableConfig } from "../investments-columns";
-import { INVESTMENT_COLUMN_MIN_SIZES } from "../net-worth-columns";
+import {
+  INVESTMENT_COLUMN_MIN_SIZES,
+  INVESTMENT_COLUMN_SIZES,
+  INVESTMENTS_TABLE_KEY,
+  investmentsTableWidth,
+  loadInvestmentsTableConfig,
+  saveInvestmentsTableConfig,
+} from "../investments-columns";
 
 beforeAll(() => installJsdomPolyfills());
 beforeEach(() => {
@@ -54,5 +62,36 @@ describe("loadInvestmentsTableConfig()", () => {
       visibility: { cost: false, pnl: true, allocation: false },
       sizing: {},
     });
+  });
+});
+
+describe("column sizes", () => {
+  it("should keep every default at or above its column's minimum", () => {
+    for (const [id, min] of Object.entries(INVESTMENT_COLUMN_MIN_SIZES)) {
+      expect(INVESTMENT_COLUMN_SIZES[id]).toBeGreaterThanOrEqual(min);
+    }
+  });
+});
+
+describe("investmentsTableWidth()", () => {
+  it("should span the same default-visible set as the account tables", () => {
+    // commodity 350 + quantity 140 + price 170 + value 170 = 830 ≈ 832.
+    expect(investmentsTableWidth(loadInvestmentsTableConfig())).toBe(830);
+  });
+
+  it("should grow when the optional Cost, P&L, and Allocation columns come on", () => {
+    expect(investmentsTableWidth({ visibility: { cost: true, pnl: true, allocation: true }, sizing: {} })).toBe(
+      830 + 170 + 170 + 120,
+    );
+  });
+
+  it("should prefer resized widths and clamp to minimums like the account tables", () => {
+    // commodity 60 clamps to its 140 minimum; the rest keep their defaults.
+    expect(
+      investmentsTableWidth({
+        visibility: { cost: false, pnl: false, allocation: false },
+        sizing: { commodity: 60 },
+      }),
+    ).toBe(INVESTMENT_COLUMN_MIN_SIZES.commodity + 140 + 170 + 170);
   });
 });
