+++ packages/desktop/src/renderer/components/accountant24/__tests__/net-worth-badge.test.tsx
@@ -43,7 +43,6 @@ const sheet = (net: NetWorth["net"], baseCommodity: NetWorth["baseCommodity"] = 
   sections: [],
   net,
   baseCommodity,
-  investments: { rows: [], totalMarketValue: [], totalCostBasis: [] },
 });
 
 const CONVERTED = sheet({
